+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Master list of UniV3 forks:
   12. ZebraV3
   13. Lynex (Algebra-like)
   14. DackieSwapV3
+  15. Thick
 
 ---
 
@@ -22,6 +23,7 @@ Master list of UniV3 forks:
 * Link to 0x's Immunefi bug bounty page from `README.md`
 * Add UniswapV3 UniV3 fork to Scroll
 * Add DackieSwapV3 UniV3 fork to Base
+* Add Thick UniV3 fork to Base
 
 ## 2024-07-29